Add initial CSS stylings to style.css
Import Lato & Russo One font.

Get rid of the margin from all sides for <body>.

Add stylings to <header>. Use Flexbox to align the logo
& the 3 header links. Also use Flexbox to put the header
links inline with a gap of 20px in between them.

Use CSS comments as headings to separate stylings for each
section of the landing page.